Move previous Key Vault related changes from settings.py to production.py
